move commands files to subfolder
